removed the previous initial timestep setting from the first iteration in the while loop and changed t to size_t.

+++ src/sph_solver.h
@@ -122,7 +122,7 @@ class SphSolver {
 
   bool adaptiveTimestepBool = false;
 
-  int t = 0;
+  size_t t = 0;
 
   int totalIterations;
